add CostedConstraintSet, no longer support setup() inits
