+++ gpkit/constraints/costed.py
@@ -1,0 +1,66 @@
+from .set import ConstraintSet
+from .prog_factories import _progify_fctry, _solve_fctry
+from ..geometric_program import GeometricProgram
+from .signomial_program import SignomialProgram
+from ..nomials import Variable
+
+
+class CostedConstraintSet(ConstraintSet):
+    def __init__(self, cost, constraints, substitutions=None):
+        self.cost = cost
+        ConstraintSet.__init__(self, constraints, substitutions)
+
+    def __getitem__(self, key):
+        if isinstance(key, int):
+            return list.__getitem__(self, key)
+        else:
+            variables = [Variable(**key.descr) for key in self.varkeys[key]]
+            if len(variables) == 1:
+                return variables[0]
+            else:
+                return variables
+    gp = _progify_fctry(GeometricProgram)
+    sp = _progify_fctry(SignomialProgram)
+    solve = _solve_fctry(_progify_fctry(GeometricProgram, "solve"))
+    localsolve = _solve_fctry(_progify_fctry(SignomialProgram, "localsolve"))
+
+    def zero_lower_unbounded_variables(self):
+        "Recursively substitutes 0 for variables that lack a lower bound"
+        zeros = True
+        while zeros:
+            bounds = self.gp(verbosity=0).missingbounds
+            zeros = {var: 0 for var, bound in bounds.items()
+                     if bound == "lower"}
+            self.substitutions.update(zeros)
+
+    def interact(self, ranges=None, fn_of_sol=None, **solvekwargs):
+        """Easy model interaction in IPython / Jupyter
+
+        By default, this creates a model with sliders for every constant
+        which prints a new solution table whenever the sliders are changed.
+
+        Arguments
+        ---------
+        fn_of_sol : function
+            The function called with the solution after each solve that
+            displays the result. By default prints a table.
+
+        ranges : dictionary {str: Slider object or tuple}
+            Determines which sliders get created. Tuple values may contain
+            two or three floats: two correspond to (min, max), while three
+            correspond to (min, step, max)
+
+        **solvekwargs
+            kwargs which get passed to the solve()/localsolve() method.
+        """
+        from ..interactive.widgets import modelinteract
+        return modelinteract(self, ranges, fn_of_sol, **solvekwargs)
+
+    def controlpanel(self, *args, **kwargs):
+        """Easy model control in IPython / Jupyter
+
+        Like interact(), but with the ability to control sliders and their
+        ranges live. args and kwargs are passed on to interact()
+        """
+        from ..interactive.widgets import modelcontrolpanel
+        return modelcontrolpanel(self, *args, **kwargs)
